% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{FUNLDA}
\alias{FUNLDA}
\alias{FUNLDA-package}
\title{FUNLDA - a latent Dirichlet allocation package for genomic annotations}
\description{
FUNLDA - a latent Dirichlet allocation package for genomic annotations
}
