% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissuenew}
\alias{tissuenew}
\title{Data for a tissue not used to to fit the LDA model}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 200 rows and 7 columns.}
\usage{
tissuenew
}
\description{
A dataset containing genomic annotations for 200 variants 
for a tissue different than the 24 tissues variants for which
are included in training
}
\keyword{datasets}
