% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training}
\alias{training}
\title{Toy training data}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 4800 rows and 7 columns.}
\usage{
training
}
\description{
A dataset containing genomic annotations for 4800 variants, 200
variants for each of 24 tissues
}
\keyword{datasets}
